+++ services/workflows-service/src/project/project-scope.service.ts
@@ -1,7 +1,6 @@
 import { Prisma } from '@prisma/client';
-import type { TProjectId, TProjectIds } from '@/types';
+import type { TProjectIds } from '@/types';
 import { Injectable } from '@nestjs/common';
-import { isUpdateOrCreateByAssociation } from '@/common/utils/prisma-helpers';
 
 export interface PrismaGeneralQueryArgs {
   select?: Record<string, unknown> | null;
@@ -65,67 +64,6 @@ export class ProjectScopeService {
       project: {
         id: {
           in: projectIds,
-        },
-      },
-    };
-
-    return args;
-  }
-
-  scopeCreate<T>(args: Prisma.SelectSubset<T, PrismaGeneralInsertArgs>, projectId?: TProjectId) {
-    // @ts-expect-error - dynamically typed for all queries
-    if (isUpdateOrCreateByAssociation(args.data)) {
-      // @ts-expect-error - dynamically typed for all queries
-      args.data = {
-        // @ts-expect-error - dynamically typed for all queries
-        ...args.data,
-        project: {
-          connect: { id: projectId },
-        },
-      };
-    } else {
-      // @ts-expect-error - dynamically typed for all queries
-      args.data = {
-        // @ts-expect-error - dynamically typed for all queries
-        ...args.data,
-        projectId,
-      };
-    }
-    return args;
-  }
-
-  scopeUpdate<T>(args: Prisma.SelectSubset<T, Prisma.FilterUpdateArgs>, projectId: TProjectId) {
-    args = this.scopeCreate(args, projectId);
-
-    return args;
-  }
-
-  scopeUpsert<T>(
-    args: Prisma.SelectSubset<T, PrismaGeneralUpsertArgs>,
-    projectIds: TProjectIds,
-    projectId: TProjectId,
-  ) {
-    // @ts-expect-error - dynamically typed for all queries
-    args.where = {
-      // @ts-expect-error - dynamically typed for all queries
-      ...args.where,
-      project: {
-        id: { in: projectIds },
-      },
-    };
-    // @ts-expect-error - dynamically typed for all queries
-    args.update = {
-      // @ts-expect-error - dynamically typed for all queries
-      ...args.update,
-      projectId: projectId,
-    };
-    // @ts-expect-error - dynamically typed for all queries
-    args.create = {
-      // @ts-expect-error - dynamically typed for all queries
-      ...args.create,
-      project: {
-        connect: {
-          id: projectId,
         },
       },
     };