+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -29,12 +29,12 @@ import {
   ChildPluginCallbackOutput,
   ChildToParentCallback,
   ChildWorkflowCallback,
-  SerializableTransformer,
-  THelperFormatingLogic,
-  Transformer,
   createWorkflow,
   HelpersTransformer,
   JmespathTransformer,
+  SerializableTransformer,
+  THelperFormatingLogic,
+  Transformer,
 } from '@ballerine/workflow-core';
 import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import {
@@ -148,7 +148,7 @@ export class WorkflowService {
       return await this.workflowDefinitionRepository.createUnscoped({ data, select });
     }
 
-    return await this.workflowDefinitionRepository.create({ data, select }, projectId);
+    return await this.workflowDefinitionRepository.create({ data, select });
   }
 
   async cloneWorkflowDefinition(data: WorkflowDefinitionCloneDto, projectId: string) {
@@ -172,14 +172,11 @@ export class WorkflowService {
       { select },
     );
 
-    return await this.workflowDefinitionRepository.create(
-      {
-        // @ts-expect-error - types of workflow definition does not propagate to the prisma creation type
-        data: { ...workflowDefinition, name: data.name, projectId: projectId, isPublic: false },
-        select,
-      },
-      [projectId],
-    );
+    return await this.workflowDefinitionRepository.create({
+      // @ts-expect-error - types of workflow definition does not propagate to the prisma creation type
+      data: { ...workflowDefinition, name: data.name, projectId: projectId, isPublic: false },
+      select,
+    });
   }
 
   async getWorkflowRuntimeDataById(
@@ -876,16 +873,13 @@ export class WorkflowService {
 
       if (allDocumentsResolved) {
         updatedWorkflow.status = allDocumentsResolved ? 'completed' : updatedWorkflow.status;
-        await this.workflowRuntimeDataRepository.updateById(
-          workflowId,
-          {
-            data: {
-              status: updatedWorkflow.status,
-              resolvedAt: new Date().toISOString(),
-            },
+        await this.workflowRuntimeDataRepository.updateById(workflowId, {
+          data: {
+            status: updatedWorkflow.status,
+            resolvedAt: new Date().toISOString(),
+            projectId,
           },
-          projectId,
-        );
+        });
 
         this.workflowEventEmitter.emit('workflow.completed', {
           runtimeData: updatedWorkflow,
@@ -993,7 +987,7 @@ export class WorkflowService {
     context: WorkflowRuntimeData['context'],
     projectId: TProjectId,
   ) {
-    return this.workflowRuntimeDataRepository.updateById(id, { data: { context } }, projectId);
+    return this.workflowRuntimeDataRepository.updateById(id, { data: { context, projectId } });
   }
 
   async updateWorkflowRuntimeData(
@@ -1099,57 +1093,48 @@ export class WorkflowService {
       );
 
       // Updates the collect documents workflow with the manual review workflow's decision.
-      await this.workflowRuntimeDataRepository.updateById(
-        parentMachine?.id,
-        {
-          data: {
-            status: 'active',
-            // @ts-expect-error - error from Prisma types fix
-            state: parentMachine?.workflowDefinition?.definition?.initial as string,
-            context: {
-              ...parentMachine?.context,
-              documents: parentMachine?.context?.documents?.map((document: any) => {
-                if (document.id !== documentToRevise.id) return document;
+      await this.workflowRuntimeDataRepository.updateById(parentMachine?.id, {
+        data: {
+          status: 'active',
+          // @ts-expect-error - error from Prisma types fix
+          state: parentMachine?.workflowDefinition?.definition?.initial as string,
+          context: {
+            ...parentMachine?.context,
+            documents: parentMachine?.context?.documents?.map((document: any) => {
+              if (document.id !== documentToRevise.id) return document;
 
-                return {
-                  ...document,
-                  decision: documentToRevise.decision,
-                };
-              }),
+              return {
+                ...document,
+                decision: documentToRevise.decision,
+              };
+            }),
+          },
+          projectId,
+        },
+      });
+
+      updatedResult = await this.workflowRuntimeDataRepository.updateById(workflowRuntimeId, {
+        data: {
+          ...data,
+          context: {
+            ...data.context,
+            parentMachine: {
+              id: parentMachine?.id,
+              status: 'active',
             },
           },
+          resolvedAt: isResolved ? new Date().toISOString() : null,
+          projectId,
         },
-        projectId,
-      );
-
-      updatedResult = await this.workflowRuntimeDataRepository.updateById(
-        workflowRuntimeId,
-        {
-          data: {
-            ...data,
-            context: {
-              ...data.context,
-              parentMachine: {
-                id: parentMachine?.id,
-                status: 'active',
-              },
-            },
-            resolvedAt: isResolved ? new Date().toISOString() : null,
-          },
-        },
-        projectId,
-      );
+      });
     } else {
-      updatedResult = await this.workflowRuntimeDataRepository.updateById(
-        workflowRuntimeId,
-        {
-          data: {
-            ...data,
-            resolvedAt: isResolved ? new Date().toISOString() : null,
-          },
+      updatedResult = await this.workflowRuntimeDataRepository.updateById(workflowRuntimeId, {
+        data: {
+          ...data,
+          resolvedAt: isResolved ? new Date().toISOString() : null,
+          projectId,
         },
-        projectId,
-      );
+      });
     }
 
     if (isResolved) {
@@ -1214,8 +1199,7 @@ export class WorkflowService {
 
     const updatedWorkflowRuntimeData = await this.workflowRuntimeDataRepository.updateById(
       workflowRuntimeId,
-      { data: { assigneeId, assignedAt: new Date() } },
-      currentProjectId,
+      { data: { assigneeId, assignedAt: new Date(), projectId: currentProjectId } },
     );
 
     if (
@@ -1285,25 +1269,17 @@ export class WorkflowService {
     const endUserId = runtime.endUserId;
     const businessId = runtime.businessId;
     endUserId &&
-      (await this.endUserRepository.updateById(
-        endUserId,
-        {
-          data: {
-            approvalState: ApprovalState.PROCESSING,
-          },
+      (await this.endUserRepository.updateById(endUserId, {
+        data: {
+          approvalState: ApprovalState.PROCESSING,
         },
-        currentProjectId,
-      ));
+      }));
     businessId &&
-      (await this.businessRepository.updateById(
-        businessId,
-        {
-          data: {
-            approvalState: ApprovalState.PROCESSING,
-          },
+      (await this.businessRepository.updateById(businessId, {
+        data: {
+          approvalState: ApprovalState.PROCESSING,
         },
-        currentProjectId,
-      ));
+      }));
 
     const entityId = endUserId || businessId;
 
@@ -1338,24 +1314,22 @@ export class WorkflowService {
     );
 
     if (!manualReviewWorkflow) {
-      await this.workflowRuntimeDataRepository.create(
-        {
-          data: {
-            ...entitySearch,
-            workflowDefinitionVersion: workflow.version,
-            workflowDefinitionId: workflow.reviewMachineId,
-            context: {
-              ...context,
-              parentMachine: {
-                id: runtime.id,
-                status: 'completed',
-              },
+      await this.workflowRuntimeDataRepository.create({
+        data: {
+          ...entitySearch,
+          workflowDefinitionVersion: workflow.version,
+          workflowDefinitionId: workflow.reviewMachineId,
+          context: {
+            ...context,
+            parentMachine: {
+              id: runtime.id,
+              status: 'completed',
             },
-            status: 'active',
           },
+          status: 'active',
+          projectId: currentProjectId,
         },
-        currentProjectId,
-      );
+      });
     } else {
       if (manualReviewWorkflow.state === 'revision') {
         await this.event(
@@ -1368,21 +1342,18 @@ export class WorkflowService {
         );
       }
 
-      await this.workflowRuntimeDataRepository.updateById(
-        manualReviewWorkflow.id,
-        {
-          data: {
-            context: {
-              ...manualReviewWorkflow.context,
-              parentMachine: {
-                id: runtime.id,
-                status: 'completed',
-              },
+      await this.workflowRuntimeDataRepository.updateById(manualReviewWorkflow.id, {
+        data: {
+          context: {
+            ...manualReviewWorkflow.context,
+            parentMachine: {
+              id: runtime.id,
+              status: 'completed',
             },
           },
+          projectId: currentProjectId,
         },
-        currentProjectId,
-      );
+      });
     }
 
     await this.updateWorkflowRuntimeData(
@@ -1524,32 +1495,30 @@ export class WorkflowService {
         customer.name,
       );
 
-      workflowRuntimeData = await this.workflowRuntimeDataRepository.create(
-        {
-          data: {
-            ...entityConnect,
-            workflowDefinitionVersion: workflowDefinition.version,
-            context: {
-              ...contextToInsert,
-              documents: documentsWithPersistedImages,
-            } as InputJsonValue,
-            config: mergedConfig as InputJsonValue,
-            // @ts-expect-error - error from Prisma types fix
-            state: workflowDefinition.definition.initial as string,
-            status: 'active',
-            workflowDefinitionId: workflowDefinition.id,
-            ...(parentWorkflowId &&
-              ({
-                parentRuntimeDataId: parentWorkflowId,
-              } satisfies Omit<
-                Prisma.WorkflowRuntimeDataCreateArgs['data'],
-                'context' | 'workflowDefinitionVersion'
-              >)),
-            ...('salesforceObjectName' in salesforceData && salesforceData),
-          },
+      workflowRuntimeData = await this.workflowRuntimeDataRepository.create({
+        data: {
+          ...entityConnect,
+          workflowDefinitionVersion: workflowDefinition.version,
+          context: {
+            ...contextToInsert,
+            documents: documentsWithPersistedImages,
+          } as InputJsonValue,
+          config: mergedConfig as InputJsonValue,
+          // @ts-expect-error - error from Prisma types fix
+          state: workflowDefinition.definition.initial as string,
+          status: 'active',
+          workflowDefinitionId: workflowDefinition.id,
+          ...(parentWorkflowId &&
+            ({
+              parentRuntimeDataId: parentWorkflowId,
+            } satisfies Omit<
+              Prisma.WorkflowRuntimeDataCreateArgs['data'],
+              'context' | 'workflowDefinitionVersion'
+            >)),
+          ...('salesforceObjectName' in salesforceData && salesforceData),
+          projectId: currentProjectId,
         },
-        currentProjectId,
-      );
+      });
 
       logDocumentWithoutId({
         line: 'createOrUpdateWorkflow 1476',
@@ -1591,9 +1560,9 @@ export class WorkflowService {
                 ...workflowRuntimeData.context,
                 metadata: { customerName: customer.displayName, token: workflowToken.token },
               } as InputJsonValue,
+              projectId: currentProjectId,
             },
           },
-          currentProjectId,
         );
       }
 
@@ -1654,9 +1623,9 @@ export class WorkflowService {
               existingWorkflowRuntimeData.config,
               validatedConfig || {},
             ) as InputJsonValue,
+            projectId: currentProjectId,
           },
         },
-        currentProjectId,
       );
 
       logDocumentWithoutId({
@@ -1738,15 +1707,13 @@ export class WorkflowService {
     context: DefaultContextSchema,
     projectId: TProjectId,
   ) {
-    const { id } = await this.endUserRepository.create(
-      {
-        data: {
-          correlationId: entity.id,
-          ...(context.entity.data as object),
-        } as Prisma.EndUserCreateInput,
-      },
-      projectId,
-    );
+    const { id } = await this.endUserRepository.create({
+      data: {
+        correlationId: entity.id,
+        ...(context.entity.data as object),
+        project: { connect: { id: projectId } },
+      } as Prisma.EndUserCreateInput,
+    });
     return id;
   }
 
@@ -1756,15 +1723,13 @@ export class WorkflowService {
     projectIds: TProjectIds,
     currentProjectId: TProjectId,
   ) {
-    const { id } = await this.businessRepository.create(
-      {
-        data: {
-          correlationId: entity.id,
-          ...(context.entity.data as object),
-        } as Prisma.BusinessCreateInput,
-      },
-      currentProjectId,
-    );
+    const { id } = await this.businessRepository.create({
+      data: {
+        correlationId: entity.id,
+        ...(context.entity.data as object),
+        project: { connect: { id: currentProjectId } },
+      } as Prisma.BusinessCreateInput,
+    });
 
     return id;
   }
@@ -1947,27 +1912,19 @@ export class WorkflowService {
     }
 
     if (entityType === 'endUser') {
-      await this.entityRepository[entityType].updateById(
-        entityId,
-        {
-          data: {
-            approvalState,
-          },
+      await this.entityRepository[entityType].updateById(entityId, {
+        data: {
+          approvalState,
         },
-        currentProjectId,
-      );
+      });
     }
 
     if (entityType === 'business') {
-      await this.entityRepository[entityType].updateById(
-        entityId,
-        {
-          data: {
-            approvalState,
-          },
+      await this.entityRepository[entityType].updateById(entityId, {
+        data: {
+          approvalState,
         },
-        currentProjectId,
-      );
+      });
     }
 
     return updatedRuntimeData;
@@ -2178,7 +2135,6 @@ export class WorkflowService {
     };
   }) {
     return await this.salesforceService.updateRecord({
-      // @ts-expect-error - error from Prisma types fix
       projectId: workflowRuntimeData.projectId,
       // @ts-expect-error - error from Prisma types fix
       objectName: workflowRuntimeData.salesforceObjectName,