+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### Patch Changes
 
-- Updated dependencies
-  - @ballerine/common@0.7.24
+- d5cca1e8: added defintion validation
+- Updated dependencies [d5cca1e8]
+- Updated dependencies [f4b5c0df]
   - @ballerine/workflow-core@0.5.23
+  - @ballerine/common@0.7.24
 
 ## 0.5.22